requirements.txt
Django==5.0.6
djangorestframework==3.15.1
psycopg2-binary==2.9.9
redis==5.0.4
django-redis==5.4.0
channels==4.0.0
channels-redis==4.2.0
celery==5.3.6
gunicorn==21.2.0
django-cors-headers==4.3.1
django-ratelimit==4.1.0
Pillow==10.3.0
python-decouple==3.8
whitenoise==6.6.0
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
django-modeltranslation==0.18.11
django-admin-interface==0.28.6
django-colorfield==0.11.0
django-storages==1.14.2
python-slugify==8.0.4
google-cloud-storage==2.11.0


